Refactoring module main.py - 13: Changed the structure of the project.
