+++ modules/sale_logic.py
@@ -2,7 +2,7 @@ import logging
 from datetime import date
 from decimal import Decimal
 
-from system import System
+from modules.system import System
 
 system = System()
